cloud-functions/terrain-generator/requirements.txt
flask==3.0.0
gunicorn==21.2.0
google-cloud-storage==2.14.0
google-cloud-firestore==2.14.0
aiohttp==3.9.0
numpy<2
shapely==2.0.0
pyshp==2.3.0
